feat(jsx/dom): introduce jsx/dom/tiny. Some features are omit in this runtime, but the size is very small.
