+++ src/jsx/dom/jsx-dev-runtime.ts
@@ -3,20 +3,5 @@
  * This module provides the `hono/jsx/dom` dev runtime.
  */
 
-import type { JSXNode, Props } from '../base'
-import * as intrinsicElementTags from './intrinsic-element/components'
-
-export const jsxDEV = (tag: string | Function, props: Props, key?: string): JSXNode => {
-  if (typeof tag === 'string' && intrinsicElementTags[tag as keyof typeof intrinsicElementTags]) {
-    tag = intrinsicElementTags[tag as keyof typeof intrinsicElementTags]
-  }
-  return {
-    tag,
-    type: tag,
-    props,
-    key,
-    ref: props.ref,
-  } as JSXNode
-}
-
-export const Fragment = (props: Record<string, unknown>): JSXNode => jsxDEV('', props, undefined)
+import './feat/intrinsic-element'
+export { jsxDEV, Fragment } from './tiny/jsx-dev-runtime'
