+++ src/jsx/dom/tiny/jsx-runtime.ts
@@ -1,0 +1,7 @@
+/**
+ * @module
+ * This module provides the `hono/jsx/dom/tiny` runtime.
+ */
+
+export { jsxDEV as jsx, Fragment } from './jsx-dev-runtime'
+export { jsxDEV as jsxs } from './jsx-dev-runtime'
