+++ src/jsx/hooks/dom.test.tsx
@@ -1,5 +1,6 @@
 /** @jsxImportSource ../ */
 import { JSDOM } from 'jsdom'
+import '../dom/feat/intrinsic-element'
 // run tests by old style jsx default
 // hono/jsx/jsx-runtime and hono/jsx/dom/jsx-runtime are tested in their respective settings
 import { ErrorBoundary, Suspense, render } from '../dom'
